Added check for negative homologies, and fixed bug in print
The integers in bmatrix was printed as int16, leading to overflow showing some homologies as negative. Changed printout to int64 to fix

+++ Python/modules/drawGraph.py
@@ -40,7 +40,7 @@ def bmatrix(a):
     return '\n'.join(rv)
 
 def drawNodes(M,x,y,scale=1):
-    node =r'\node (' + M.id +') at (' + str(x)+ ','+ str(y)+') [scale='+str(scale)+'] '+r'{$'+bmatrix(M.homologies[::].astype(np.int16))+r'$};'+'\n'
+    node =r'\node (' + M.id +') at (' + str(x)+ ','+ str(y)+') [scale='+str(scale)+'] '+r'{$'+bmatrix(M.homologies[::].astype(np.int64))+r'$};'+'\n'
     return node
 
 def drawIrrArrows(M):
